chore: remove FLAG_CC_SOLEUR_GO (always-on in prd and dev)
Closes #3270

- Drop command-center-soleur-go entry from FLAG_VARS in
  apps/web-platform/lib/feature-flags/server.ts.
- Delete resolveInitialRouting() and its consumer in ws-handler.ts;
  initial routing is now the literal { kind: "soleur_go_pending" }.
- Drop the if (ccFlagEnabled) wrapper so the cc rate limiter runs
  unconditionally for every start_session.
- Reframe router-flag-stickiness.test.ts -> router-stickiness-
  invariant.test.ts: assert the active_workflow -> ConversationRouting
  invariant directly, without flag input.
- Update stale comments in cc-dispatcher.ts and the ADR-022 status
  sentence to reflect Stage 8 landing.
- Remove FLAG_CC_SOLEUR_GO from apps/web-platform/.env.example.

Doppler env-var unset across prd/dev/ci is post-merge operator work
(no live reader after this lands).
