+++ apps/web-platform/.env.example
@@ -78,13 +78,6 @@ NEXT_PUBLIC_SENTRY_DSN=
 # Only "1" is truthy; any other value (including unset) means off.
 FLAG_KB_CHAT_SIDEBAR=0
 
-# Route Command Center conversations via `/soleur:go` (new single-leader
-# runner) instead of the legacy multi-leader domain-router. Read ONCE at
-# `start_session`; a mid-conversation flag flip does NOT re-route in-flight
-# conversations (router-flag-stickiness). See
-# knowledge-base/project/plans/2026-04-23-feat-cc-route-via-soleur-go-plan.md.
-FLAG_CC_SOLEUR_GO=0
-
 # --- Dev-only — DO NOT SET IN PRD ---
 # Enables the multi-account QA sign-in panel on /login (R3 / feat-dev-
 # signin-bypass). The panel + POST /api/auth/dev-signin route are gated
