+++ apps/web-platform/lib/feature-flags/server.test.ts
@@ -49,52 +49,21 @@ describe("getFeatureFlags", () => {
 
   it("returns all flags as a record", () => {
     process.env.FLAG_KB_CHAT_SIDEBAR = "1";
-    delete process.env.FLAG_CC_SOLEUR_GO;
     delete process.env.FLAG_DEV_SIGNIN;
     const flags = getFeatureFlags();
     expect(flags).toEqual({
       "kb-chat-sidebar": true,
-      "command-center-soleur-go": false,
       "dev-signin": false,
     });
   });
 
   it("returns false for all flags when none are set", () => {
     delete process.env.FLAG_KB_CHAT_SIDEBAR;
-    delete process.env.FLAG_CC_SOLEUR_GO;
     delete process.env.FLAG_DEV_SIGNIN;
     const flags = getFeatureFlags();
     expect(flags).toEqual({
       "kb-chat-sidebar": false,
-      "command-center-soleur-go": false,
       "dev-signin": false,
     });
-  });
-});
-
-describe("command-center-soleur-go flag", () => {
-  const ORIGINAL_ENV = process.env;
-
-  beforeEach(() => {
-    process.env = { ...ORIGINAL_ENV };
-  });
-
-  afterEach(() => {
-    process.env = ORIGINAL_ENV;
-  });
-
-  it("reads FLAG_CC_SOLEUR_GO=1 as true", () => {
-    process.env.FLAG_CC_SOLEUR_GO = "1";
-    expect(getFlag("command-center-soleur-go")).toBe(true);
-  });
-
-  it("reads FLAG_CC_SOLEUR_GO=0 as false", () => {
-    process.env.FLAG_CC_SOLEUR_GO = "0";
-    expect(getFlag("command-center-soleur-go")).toBe(false);
-  });
-
-  it("treats unset FLAG_CC_SOLEUR_GO as false (safe default)", () => {
-    delete process.env.FLAG_CC_SOLEUR_GO;
-    expect(getFlag("command-center-soleur-go")).toBe(false);
   });
 });
