+++ apps/web-platform/lib/feature-flags/server.ts
@@ -9,7 +9,6 @@
 
 const FLAG_VARS = {
   "kb-chat-sidebar": "FLAG_KB_CHAT_SIDEBAR",
-  "command-center-soleur-go": "FLAG_CC_SOLEUR_GO",
   "dev-signin": "FLAG_DEV_SIGNIN",
 } as const;
