+++ knowledge-base/project/specs/feat-one-shot-3270-remove-flag-cc-soleur-go/session-state.md
@@ -1,0 +1,21 @@
+# Session State
+
+## Plan Phase
+- Plan file: /home/jean/git-repositories/jikig-ai/soleur/.worktrees/feat-one-shot-3270-remove-flag-cc-soleur-go/knowledge-base/project/plans/2026-05-11-chore-remove-flag-cc-soleur-go-plan.md
+- Status: complete
+
+### Errors
+None.
+
+### Decisions
+- Issue body line citation `cc-dispatcher.ts:375` is wrong — actual stale comment with `FLAG_CC_SOLEUR_GO` is at line 419. Plan targets lines 6-10 and 419.
+- Issue body `ws-handler.ts:676-703` line numbers are stale (file evolved after #3263 / migration 032). Plan targets actual range 975-1018.
+- Added `apps/web-platform/.env.example:81-86` to Files to Edit via class-wide `git grep -F FLAG_CC_SOLEUR_GO` (not in original issue body).
+- User-Brand Impact threshold = `none`. Dead-code removal (`FLAG_CC_SOLEUR_GO=1` in both prd/dev Doppler per ADR-022); no PII/auth/payment/regulated-data surface.
+- Overlapping code-review issues (#3374, #3372, #2191, #3369, #3243, #3242, #2955) Acknowledged — different concerns, no fold-in. PR uses `Closes #3270` on its own body line only.
+- Doppler env-var unset is post-merge / out-of-scope per issue body; tracked as optional Post-merge AC.
+
+### Components Invoked
+- skill: soleur:plan (inline)
+- skill: soleur:deepen-plan (inline)
+- Bash, Read tooling
